Introduce factory methods for TemplateParameter
Also internalize all template converters to reduce API surface.
